pyproject.toml
[project]
name = "waldur-cscs-hpc-storage"
version = "0.7.0"
description = "CSCS HPC Storage proxy"
readme = "README.md"
authors = [{ name = "OpenNode Team", email = "info@opennodecloud.com" }]
requires-python = ">=3.11, <4"
dependencies = [
    "fastapi-keycloak-middleware>=1.3.0",
    "fastapi[standard]>=0.115.14",
    "httpx[socks]>=0.24.0",
    "waldur-api-client>=7.9.2",
    "pyyaml>=6.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Script entry points
[project.scripts]
waldur_cscs_storage_sync = "waldur_cscs_hpc_storage.sync_script:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "respx",
    "pytest-cov",
]
